requirements.txt
fastmcp>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
pytest>=7.0.0
uvicorn>=0.24.0 